# backend/specs/001-book-rag-chatbot/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: RAG Chatbot API
  version: 1.0.0
  description: API for the integrated RAG chatbot that answers questions about book content

paths:
  /api/v1/chat/start:
    post:
      summary: Start a new chat session
      description: Creates a new session for a user to interact with the book content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                book_id:
                  type: string
                  format: uuid
                  description: The ID of the book to interact with
                user_id:
                  type: string
                  format: uuid
                  description: Optional user ID if applicable
              required:
                - book_id
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: "Session created successfully"
        '400':
          description: Invalid book ID or other parameters
        '404':
          description: Book not found

  /api/v1/chat/query:
    post:
      summary: Submit a query to the chatbot
      description: Asks the chatbot a question about the book content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: The ID of the session
                question:
                  type: string
                  description: The question to ask
                rag_mode:
                  type: string
                  enum: [global, selection_only]
                  description: The RAG mode to use (default is global)
                selection_context:
                  type: string
                  description: Selected text for selection-only mode (required if rag_mode is selection_only)
              required:
                - session_id
                - question
      responses:
        '200':
          description: Query processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: The chatbot's answer to the question
                  citations:
                    type: array
                    items:
                      type: object
                      properties:
                        chunk_id:
                          type: string
                          description: ID of the cited chunk
                        text:
                          type: string
                          description: The cited text
                        page_number:
                          type: integer
                          description: Page number where the citation appears
                  confidence_score:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                    description: Confidence score of the answer
        '400':
          description: Invalid session ID, question, or other parameters
        '404':
          description: Session not found

  /api/v1/chat/history:
    get:
      summary: Get conversation history
      description: Retrieves the conversation history for a session
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Conversation history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    role:
                      type: string
                      enum: [user, assistant]
                    content:
                      type: string
                    timestamp:
                      type: string
                      format: date-time
                    citations:
                      type: array
                      items:
                        type: object
                        properties:
                          chunk_id:
                            type: string
                          text:
                            type: string
                          page_number:
                            type: integer
        '400':
          description: Invalid session ID
        '404':
          description: Session not found

  /api/v1/chat/end:
    post:
      summary: End a chat session
      description: Ends the specified chat session and cleans up resources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: The ID of the session to end
              required:
                - session_id
      responses:
        '200':
          description: Session ended successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session ended successfully"
        '400':
          description: Invalid session ID
        '404':
          description: Session not found

  /api/v1/books:
    get:
      summary: List available books
      description: Retrieves a list of books available for querying
      responses:
        '200':
          description: List of available books
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    book_id:
                      type: string
                      format: uuid
                    title:
                      type: string
                    author:
                      type: string
                    description:
                      type: string

  /api/v1/books/{book_id}:
    get:
      summary: Get book details
      description: Retrieves details about a specific book
      parameters:
        - name: book_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  book_id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  author:
                    type: string
                  description:
                    type: string
                  chunk_count:
                    type: integer
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '404':
          description: Book not found